requirements.txt
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.2.4
asttokens==2.4.1
async-lru==2.0.5
attrs==25.3.0
autopep8==2.3.1
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
colorgram.py==1.2.0
comm==0.2.2
cryptography==45.0.4
Cython==3.0.11
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
exif==1.6.1
fastjsonschema==2.21.1
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
norminette==3.3.51
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.1.1
objprint==0.2.3
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plum-py==0.8.7
prettytable==3.11.0
prometheus_client==0.22.1
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.18.0
pylint==3.2.7
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-networkdays==1.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==24.0.1
qrcode==8.2
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
segno==1.6.6
selectolax==0.3.30
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
terminado==0.18.1
textual==3.5.0
tinycss2==1.4.0
tomli==2.0.1
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
ttkbootstrap==1.13.11
types-python-dateutil==2.9.0.20250516
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
